pyproject.toml
[project]
name = "voice-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ollama>=0.6.1",
    "openai-whisper>=20250625",
    "streamlit>=1.56.0",
    "streamlit-webrtc>=0.64.5",
]
